Add files via upload
This repository contains Nashville housing data intended for data cleaning tasks using SQL. The dataset includes detailed information on housing properties in the Nashville area, covering various attributes such as property size, location, owner name, land value and sale prices. The goal of this project is to perform data cleaning operations using SQL queries to preprocess and prepare the dataset for further analysis or modeling. Contributions, suggestions, and collaboration are welcome to enhance the quality and usability of the dataset.